LPS-portal-service Move the method to DataSourceFactory. Since LiferayServiceExtender and ModuleApplicationContext invoke the method. The following commit will move LiferayServiceExtender to another module.
